# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "User login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrappedAuthTokenResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "summary": "Register new user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Registration successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrappedAuthTokenResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "summary": "Request password reset OTP",
        "description": "Sends a 6-digit OTP to the user's email. OTP expires after 10 minutes.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestPasswordResetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP sent if email exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrappedOkResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "summary": "Confirm password reset with OTP",
        "description": "Validates the OTP and sets the new password. All existing sessions are revoked.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmPasswordResetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrappedOkResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/verify-email": {
      "post": {
        "operationId": "AuthController_verifyEmail",
        "summary": "Verify email with OTP",
        "description": "Confirms the user's email address using the 6-digit OTP.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email verified",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrappedOkResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Hajor API",
    "description": "API documentation for Hajor fintech system",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "saliuayodeji3@gmail.com"
          },
          "password": {
            "type": "string",
            "format": "password"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "USER",
          "PROXY",
          "SUPER_ADMIN"
        ]
      },
      "AuthTokenResponseDto": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "JWT Bearer token"
          },
          "userId": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          }
        },
        "required": [
          "access_token"
        ]
      },
      "WrappedAuthTokenResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "timestamp": {
            "type": "string",
            "example": "2026-04-06T00:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/endpoint"
          },
          "requestId": {
            "type": "string",
            "nullable": true,
            "example": "uuid-here"
          },
          "data": {
            "$ref": "#/components/schemas/AuthTokenResponseDto"
          },
          "code": {
            "type": "string",
            "example": "OK"
          }
        },
        "required": [
          "statusCode",
          "timestamp",
          "path",
          "data",
          "code"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "First name",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "description": "Last name",
            "example": "Doe"
          },
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "john.doe@example.com"
          },
          "phone": {
            "type": "string",
            "description": "Nigerian phone number in international format",
            "example": "+2347038939208",
            "pattern": "^\\+234\\d{10}$"
          },
          "bvn": {
            "type": "string",
            "description": "Bank Verification Number",
            "example": "12345678901"
          },
          "dob": {
            "format": "date-time",
            "type": "string",
            "description": "Date of birth (must be at least 16 years old)",
            "example": "1995-06-15"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "StrongP@ssw0rd",
            "format": "password"
          },
          "address": {
            "type": "string",
            "description": "Residential address",
            "example": "12 Lagos Street, Abuja"
          },
          "utilityBillUrl": {
            "type": "string",
            "description": "URL of uploaded utility bill document",
            "example": "https://cdn.example.com/bills/abc.pdf"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "phone",
          "dob",
          "password"
        ]
      },
      "OkResponseDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ]
      },
      "WrappedOkResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "timestamp": {
            "type": "string",
            "example": "2026-04-06T00:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/endpoint"
          },
          "requestId": {
            "type": "string",
            "nullable": true,
            "example": "uuid-here"
          },
          "data": {
            "$ref": "#/components/schemas/OkResponseDto"
          },
          "code": {
            "type": "string",
            "example": "OK"
          }
        },
        "required": [
          "statusCode",
          "timestamp",
          "path",
          "data",
          "code"
        ]
      },
      "RequestPasswordResetDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "ConfirmPasswordResetDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "otp": {
            "type": "string",
            "example": "123456",
            "description": "6-digit OTP sent to the user"
          },
          "newPassword": {
            "type": "string",
            "example": "NewP@ssw0rd",
            "description": "New password (min 8 characters)"
          }
        },
        "required": [
          "email",
          "otp",
          "newPassword"
        ]
      },
      "VerifyEmailDto": {
        "type": "object",
        "properties": {
          "otp": {
            "type": "string",
            "description": "6-digit OTP sent to your email",
            "example": "123456"
          }
        },
        "required": [
          "otp"
        ]
      },
...
    }
  }
}